@@ -2,29 +2,34 @@
   "name": "whatrecord",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "serve": "vite preview",
-    "build": "vite build",
+    "build": "vue-tsc && vite build",
     "dev": "vite",
-    "lint": "eslint --ext .js,.vue --ignore-path ../.gitignore --fix src"
+    "lint": "eslint --ext .ts,.js,.vue --ignore-path ../.gitignore --fix src"
   },
   "dependencies": {
+    "@types/d3-graphviz": "^2.6.7",
     "axios": "^1.4.0",
-    "cytoscape": "^3.25.0",
-    "cytoscape-fcose": "^2.2.0",
+    "d3-graphviz": "^5.1.0",
+    "d3-selection": "^3.0.0",
     "es6-promise": "^4.2.8",
-    "primeflex": "^3.3.0",
+    "pinia": "^2.1.6",
+    "primeflex": "^3.3.1",
     "primeicons": "^6.0.1",
-    "primevue": "3.29.1",
+    "primevue": "^3.32.2",
     "vue": "^3.3.4",
-    "vue-router": "^4.2.1",
-    "vuex": "^4.1.0"
+    "vue-router": "^4.2.4",
+    "wildcard-match": "^5.1.2"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^4.2.3",
-    "eslint": "8",
-    "eslint-plugin-vue": "9.14.0",
-    "vite": "^4.3.8"
+    "eslint": "^8.47.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.9",
+    "vue-tsc": "^1.8.8"
   },
   "eslintConfig": {
     "root": true,
