@@ -1,53 +1,47 @@
 <template>
   <h3>{{ whatrec.name }}</h3>
   <!-- Available in {{ available_protocols }} -->
-  <template
-    v-for="([defn, instance], idx) in [
-      [record_defn, record],
-      [null, pva_group],
-    ]"
-    :key="idx"
-  >
-    <template v-if="instance != null">
+  <template v-for="pair in displayed_records">
+    <template v-if="pair.instance != null">
       <script-context-link
-        :context="instance.context"
+        :context="pair.instance.context"
         :short="0"
       ></script-context-link>
       <br />
       <epics-format-record
-        :name="instance.name"
-        :aliases="instance.aliases"
-        :context="instance.context"
-        :fields="instance.fields"
-        :record_type="instance.record_type"
-        :is_grecord="instance.is_grecord"
-        :is_pva="instance.is_pva"
-        :info_nodes="instance.info"
-        :metadata="instance.metadata"
-        :record_defn="defn"
+        :name="pair.instance.name"
+        :aliases="pair.instance.aliases"
+        :context="pair.instance.context"
+        :fields="pair.instance.fields"
+        :record_type="pair.instance.record_type"
+        :is_grecord="pair.instance.is_grecord"
+        :is_pva="pair.instance.is_pva"
+        :info_nodes="pair.instance.info"
+        :metadata="pair.instance.metadata"
+        :record_defn="pair.definition"
         :menus="whatrec.menus"
       />
 
-      <template v-for="plugin in plugins" :key="plugin.name">
-        <template
-          v-for="plugin_match in instance.metadata[plugin.name] || []"
-          :key="plugin_match"
-        >
+      <template
+        v-for="[plugin, plugin_match] of Object.entries(pair.plugin_matches)"
+        :key="plugin_match"
+      >
+        <template v-for="item of plugin_match">
           <details>
             <summary>
-              {{ plugin.name }} - {{ plugin_match.name }}
-              <router-link :to="`/${plugin.name}/${plugin_match.name}`">
+              {{ plugin }} - {{ item.name }}
+              <router-link :to="{ name: plugin, query: { item: item.name } }">
                 (Details)
               </router-link>
             </summary>
             <dictionary-table
-              :dict="plugin_match"
+              :dict="item"
               :cls="'metadata'"
               :skip_keys="['_whatrecord']"
             />
           </details>
-          <br />
         </template>
+        <br />
       </template>
     </template>
   </template>
@@ -65,7 +59,7 @@
         )
       </summary>
       <dictionary-table
-        :dict="streamdevice_metadata"
+        :dict="streamdevice_metadata as any"
         :cls="'metadata'"
         :skip_keys="['protocol']"
       />
@@ -80,7 +74,7 @@
         <br />
         <span class="code">
           <template
-            v-for="command in streamdevice_metadata.protocol.commands"
+            v-for="command in streamdevice_commands"
             :key="command.name"
           >
             {{ command.name }} {{ command.arguments.join(" ") }}<br />
@@ -92,22 +86,31 @@
   </template>
 
   <Accordion :multiple="true" class="accordion">
-    <AccordionTab :header="`Part of ${whatrec.ioc.name}`">
-      <ioc-info :ioc_info="whatrec.ioc" />
+    <AccordionTab :header="`Part of ${whatrec.ioc?.name}`">
+      <ioc-info :ioc_info="whatrec.ioc" v-if="whatrec.ioc" />
     </AccordionTab>
-    <AccordionTab header="Record links">
-      <a :href="graph_link" target="_blank">
-        <img class="svg-graph" :src="graph_link" />
-      </a>
+    <AccordionTab header="Record links" v-if="show_graph" ref="link_tab">
+      <div id="graph_div" ref="graph_div">
+        <GraphvizGraph
+          :width="graph_width"
+          :height="graph_height"
+          :fit_graph="false"
+          :save_filename="record?.name"
+          :dot_source="pv_graph_dot_source"
+        />
+      </div>
     </AccordionTab>
-    <AccordionTab header="Archiver" v-if="record != null">
+    <AccordionTab
+      header="Archiver"
+      v-if="record != null && appliance_viewer_url"
+    >
       <template v-if="record != null && appliance_viewer_url">
         <a :href="appliance_viewer_url" target="_blank"> Archive Viewer </a>
         <iframe :src="appliance_viewer_url" title="Archive viewer" />
       </template>
     </AccordionTab>
     <AccordionTab header="Autosave" v-if="record && autosave != null">
-      <template v-if="autosave?.disconnected">
+      <template v-if="autosave.disconnected">
         Disconnected fields:
         <ul>
           <li v-for="field in autosave.disconnected" :key="field">
@@ -115,7 +118,7 @@
           </li>
         </ul>
       </template>
-      <template v-for="(table, idx) in autosave_restore_tables" :key="idx">
+      <template v-for="(table, _idx) in autosave_restore_tables" :key="_idx">
         <DataTable :value="Object.values(table)" dataKey="name">
           <Column
             field="field"
@@ -132,24 +135,21 @@
         </DataTable>
       </template>
       <template v-if="autosave.error">
-        Autosave errors on {{ whatrec.ioc.name }}:
+        Autosave errors on {{ whatrec.ioc?.name }}:
         <template v-for="error of autosave.error ?? []" :key="error">
           <pre>{{ error }}</pre>
         </template>
       </template>
     </AccordionTab>
-    <AccordionTab
-      header="Gateway"
-      v-if="record?.metadata?.gateway?.matches?.length > 0"
-    >
-      <gateway-matches :matches="record.metadata.gateway.matches" />
+    <AccordionTab header="Gateway" v-if="gateway_matches">
+      <gateway-matches :matches="gateway_matches" />
     </AccordionTab>
     <AccordionTab header="Access Security Group" v-if="asg != null">
       <dictionary-table :dict="asg" :cls="'metadata'" :skip_keys="[]" />
     </AccordionTab>
-    <AccordionTab header="Asyn" v-if="record?.metadata?.asyn?.length > 0">
+    <AccordionTab header="Asyn" v-if="asyn">
       <asyn-port
-        v-for:="(asyn_port, idx) in record.metadata.asyn"
+        v-for:="asyn_port in asyn.ports"
         :asyn_port="asyn_port"
         :key="asyn_port.name"
       />
@@ -170,11 +170,14 @@
   </Accordion>
 </template>
 
-<script>
+<script lang="ts">
+import type { PropType } from "vue";
+
 import AsynPort from "./asyn-port.vue";
 import DictionaryTable from "./dictionary-table.vue";
 import EpicsFormatRecord from "./epics-format-record.vue";
 import GatewayMatches from "./gateway-matches.vue";
+import GraphvizGraph from "./graphviz-graph.vue";
 import IocInfo from "./ioc-info.vue";
 import RecordFieldTable from "./record-field-table.vue";
 import ScriptContextLink from "./script-context-link.vue";
@@ -184,12 +187,54 @@ import AccordionTab from "primevue/accordiontab";
 import Column from "primevue/column";
 import DataTable from "primevue/datatable";
 
-import { plugins } from "../settings.js";
+import { use_configured_store } from "@/stores";
+import { Plugin, plugins } from "../settings";
+import {
+  RecordType,
+  WhatRecord,
+  GatewayMatch,
+  GatewayMetadata,
+  AutosaveMetadataAnnotation,
+  RecordInstance,
+} from "../types";
+
+import { AsynMetadata } from "@/types/asyn";
+import {
+  Command as StreamDeviceCommand,
+  StreamDeviceRecordAnnotation,
+} from "@/types/stream";
+
+interface PluginMatch {
+  name: string;
+  [x: string]: any;
+}
+
+interface DisplayedRecords {
+  definition: RecordType | null;
+  instance: RecordInstance;
+  plugin_matches: Record<string, PluginMatch[]>;
+}
+
+function get_plugin_matches(
+  metadata: Record<string, any>,
+): Record<string, any> {
+  // v-for="plugin_match in pair.instance.metadata[plugin.name] || {}"
+  let result: Record<string, any> = {};
+  for (const plugin of plugins) {
+    if (plugin.name in metadata) {
+      result[plugin.name] = metadata[plugin.name];
+    }
+  }
+  return result;
+}
 
 export default {
   name: "Recordinfo",
   props: {
-    whatrec: Object,
+    whatrec: {
+      type: Object as PropType<WhatRecord>,
+      required: true,
+    },
   },
   components: {
     AsynPort,
@@ -198,25 +243,62 @@ export default {
     DictionaryTable,
     EpicsFormatRecord,
     GatewayMatches,
+    GraphvizGraph,
     IocInfo,
     RecordFieldTable,
     ScriptContextLink,
     Accordion,
     AccordionTab,
   },
+  setup() {
+    const store = use_configured_store();
+    return { store };
+  },
+  data() {
+    return {
+      pv_graph_dot_source: "",
+      graph_width: 300,
+      graph_height: 300,
+      show_graph: true,
+    };
+  },
   computed: {
+    displayed_records(): DisplayedRecords[] {
+      let result = [];
+      if (this.record != null) {
+        result.push({
+          definition: this.record_defn,
+          instance: this.record,
+          plugin_matches: get_plugin_matches(this.record.metadata),
+        });
+        console.log("matches", result);
+      }
+      if (this.pva_group != null) {
+        result.push({
+          definition: null,
+          instance: this.pva_group,
+          plugin_matches: get_plugin_matches(this.pva_group.metadata),
+        });
+      }
+      return result;
+    },
     appliance_viewer_url() {
-      const appliance_viewer_url =
-        import.meta.env.WHATRECORD_ARCHIVER_URL || "";
-      if (!appliance_viewer_url || !this.record) {
+      const url: string = import.meta.env.WHATRECORD_ARCHIVER_URL || "";
+      if (!url || !this.record) {
         return null;
       }
-      return appliance_viewer_url + this.record.name;
+      return url + this.record.name;
     },
     graph_link() {
+      if (!this.store.is_online) {
+        return null;
+      }
       return `/api/pv/graph?pv=${this.whatrec.name}&format=svg`;
     },
     script_graph_link() {
+      if (!this.store.is_online) {
+        return null;
+      }
       return `/api/pv/script-graph?pv=${this.whatrec.name}&format=svg`;
     },
     available_protocols() {
@@ -235,31 +317,81 @@ export default {
     record() {
       return this.whatrec.record ? this.whatrec.record.instance : null;
     },
-    record_defn() {
-      return this.whatrec.record ? this.whatrec.record.definition : null;
+    record_defn(): RecordType | null {
+      if (this.whatrec.record == null) {
+        return null;
+      }
+      return this.whatrec.record.definition ?? null;
     },
     asg() {
       return this.record?.metadata["asg"];
     },
-    autosave() {
-      return this.record?.metadata["autosave"] ?? null;
+    autosave(): AutosaveMetadataAnnotation | null {
+      return (
+        (this.record?.metadata["autosave"] as AutosaveMetadataAnnotation) ??
+        null
+      );
+    },
+    asyn(): AsynMetadata | null {
+      return (this.record?.metadata["asyn"] as AsynMetadata) ?? null;
+    },
+    gateway_matches(): GatewayMatch[] | null {
+      const gateway: GatewayMetadata | null =
+        (this.record?.metadata["gateway"] as GatewayMetadata) ?? null;
+      if (!gateway) {
+        return null;
+      }
+      return gateway.matches;
     },
     autosave_restore_tables() {
       return this.autosave?.restore ?? [];
     },
-    streamdevice_metadata() {
-      return this.record?.metadata["streamdevice"];
+    streamdevice_metadata(): StreamDeviceRecordAnnotation | null {
+      // this shows I don't know what I'm doing with typescript except silencing
+      // the linter:
+      let result = this.record?.metadata?.streamdevice ?? null;
+      return result ? (result as StreamDeviceRecordAnnotation) : null;
     },
-    plugins() {
+    streamdevice_commands(): StreamDeviceCommand[] {
+      return this.streamdevice_metadata?.protocol?.commands ?? [];
+    },
+    plugins(): Plugin[] {
       return plugins;
+    },
+  },
+
+  async mounted() {
+    await this.update_record_info();
+  },
+
+  async updated() {
+    await this.update_record_info();
+  },
+
+  methods: {
+    async update_record_info() {
+      const graph = await this.store.get_record_link_graph({
+        record_name: this.whatrec.name,
+      });
+      const graph_div = this.$refs.graph_div as HTMLDivElement | null;
+      const link_tab = this.$refs.link_tab as HTMLElement | null;
+      if (!graph_div || !link_tab) {
+        console.error("What happened? No graph or tab?");
+        return;
+      }
+      this.graph_width = graph_div.clientWidth || window.innerWidth * 0.75;
+      this.graph_height = graph_div.clientHeight || window.innerHeight * 0.25;
+      this.pv_graph_dot_source = graph?.source ?? "";
+      this.show_graph = this.pv_graph_dot_source.length > 0;
     },
   },
 };
 </script>
 
 <style scoped>
-.svg-graph {
-  max-width: 70%;
+#button-save-svg {
+  max-width: 10%;
+  max-height: 10%;
 }
 
 .monospace {
@@ -293,5 +425,12 @@ iframe {
   padding: 15px;
   page-break-inside: avoid;
   word-wrap: break-word;
+}
+
+#graph_div {
+  height: 100%;
+  width: 100%;
+  min-width: 50%;
+  min-height: 50%;
 }
 </style>
