@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 import { router } from "./router";
-import { store } from "./store";
+import { createPinia } from "pinia";
 
 import App from "./App.vue";
 import PrimeVue from "primevue/config";
@@ -10,8 +10,10 @@ import "primevue/resources/primevue.min.css";
 import "primeflex/primeflex.css";
 import "primeicons/primeicons.css";
 
+const pinia = createPinia();
 const app = createApp(App);
+
 app.use(router);
 app.use(PrimeVue);
-app.use(store);
+app.use(pinia);
 app.mount("#app");
