@@ -6,7 +6,14 @@ export const netconfig_enabled = enabled_plugins.indexOf("netconfig") >= 0;
 export const twincat_pytmc_enabled =
   enabled_plugins.indexOf("twincat_pytmc") >= 0;
 export const epicsarch_enabled = enabled_plugins.indexOf("epicsarch") >= 0;
-let _plugin_info = [
+
+export interface Plugin {
+  name: string;
+  label: string;
+  icon: string;
+}
+
+let _plugin_info: Plugin[] = [
   {
     name: "happi",
     label: "happi",
@@ -29,6 +36,6 @@ let _plugin_info = [
   },
 ];
 
-export const plugins = _plugin_info.filter(
-  (plugin) => enabled_plugins.indexOf(plugin.name) >= 0
+export const plugins: Plugin[] = _plugin_info.filter(
+  (plugin) => enabled_plugins.indexOf(plugin.name) >= 0,
 );
