@@ -2,16 +2,14 @@ import asyncio
 import collections
 import fnmatch
 import functools
+import gzip
+import json
 import logging
 import os
+import pathlib
 import re
 import tempfile
-
-try:
-    import tracemalloc
-except ImportError:
-    # tracemalloc unavailable on pypy
-    tracemalloc = None
+import time
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 import apischema
@@ -19,15 +17,22 @@ import graphviz
 from aiohttp import web
 
 from .. import common, gateway, graph, ioc_finder, settings
-from ..common import LoadContext, RecordInstance, StringWithContext, WhatRecord
-from ..shell import (LoadedIoc, ScriptContainer, ShellState,
+from ..common import (AnyPath, LoadContext, RecordInstance, StringWithContext,
+                      WhatRecord)
+from ..shell import (IocshScript, LoadedIoc, ScriptContainer, ShellState,
                      load_startup_scripts_with_metadata)
 from .common import (IocGetDuplicatesResponse, IocGetMatchesResponse,
-                     IocGetMatchingRecordsResponse, IocMetadata, PVGetInfo,
-                     PVGetMatchesResponse, PVRelationshipResponse,
+                     IocGetMatchingRecordsResponse, IocMetadata, PluginResults,
+                     PVGetInfo, PVGetMatchesResponse, PVRelationshipResponse,
                      PVShortRelationshipResponse, ServerPluginSpec,
                      TooManyRecordsError)
 from .util import TaskHandler
+
+try:
+    import tracemalloc
+except ImportError:
+    # tracemalloc unavailable on pypy
+    tracemalloc = None
 
 TRUE_VALUES = {"1", "true", "True"}
 
@@ -112,6 +117,222 @@ class ServerState:
         ]
         self._update_count = 0
 
+    async def trigger_manual_update(self, annotate_records: bool = True) -> None:
+        """Manual re-scan of everything."""
+        updated = await self._update_script_step()
+        logger.debug(
+            "Manual update trigger. Updates %s",
+            "found" if updated else "not found",
+        )
+        for plugin in self.plugins:
+            await self._update_plugin(plugin)
+
+        # self.container.pv_relations = graph.build_database_relations(
+        #     database=self.container.database,
+        #     record_types=self.container.record_types,  # <-- inaccurate?
+        #     aliases=self.container.aliases,
+        # )
+        self.script_relations = graph.build_script_relations(
+            self.container.database,
+            self.container.pv_relations,
+        )
+
+        if not annotate_records:
+            return
+
+        for ioc in self.container.scripts.values():
+            logger.info(
+                "Annotating %d records for IOC %s",
+                len(ioc.shell_state.database), ioc.name,
+            )
+            for record in ioc.shell_state.database.values():
+                record.metadata.update(self.get_record_annnotations(ioc, record))
+                for fn, hash_ in ioc.shell_state.loaded_files.items():
+                    if fn not in self.container.loaded_files:
+                        logger.info(
+                            "Record annotation for %s loaded a new file dynamically: %s (%s)",
+                            ioc.name,
+                            fn,
+                            hash_,
+                        )
+                        self.container.loaded_files[fn] = hash_
+
+    def dump_file(self, filename: AnyPath, recorded_hash: str = "") -> dict:
+        # TODO recorded hash may differ from what's on disk!
+        ioc_name = self.container.startup_script_to_ioc.get(str(filename), None)
+        if ioc_name:
+            logger.debug("File %s is from IOC %s", filename, ioc_name)
+            loaded_ioc = self.container.scripts[ioc_name]
+            script_info = loaded_ioc.script
+            ioc_md = loaded_ioc.metadata
+            return {
+                "script": apischema.serialize(IocshScript, script_info),
+                "ioc": apischema.serialize(IocMetadata, ioc_md),
+            }
+
+        try:
+            with open(filename, "rt") as fp:
+                lines = fp.read().splitlines()
+        except Exception as ex:
+            lines = [
+                f"whatrecord unable to dump file due to "
+                f"{ex.__class__.__name__}: {ex}"
+            ]
+
+        # TODO: make these into serialized dataclass responses
+        return {
+            "script": None,
+            "ioc": None,
+            "raw": {
+                "lines": lines,
+                "path": str(filename),
+                "hash": recorded_hash,
+            },
+        }
+
+    def dump_iocs(self) -> dict:
+        return {
+            ioc.metadata.name: {
+                "md": apischema.serialize(IocMetadata, ioc.metadata),
+                "ioc": apischema.serialize(LoadedIoc, ioc),
+            }
+            for ioc in self.container.scripts.values()
+            if ioc.metadata and ioc.metadata.name
+        }
+
+    def dump_plugins(self) -> dict:
+        plugins = {
+            plugin.name: apischema.serialize(
+                Optional[PluginResults],
+                plugin.results,
+            )
+            for plugin in self.plugins
+        }
+        plugins["gateway"] = apischema.serialize(
+            PluginResults,
+            PluginResults(
+                metadata=self.gateway_config,
+            )
+        )
+        return plugins
+
+    def dump_files(self, *, exclude_extensions: Optional[list[str]]) -> dict:
+        exclude_extensions = exclude_extensions or []
+        result = {}
+        hash_to_contents = result["hash_to_contents"] = {}
+        hash_to_file = result["hash_to_file"] = {}
+        for fn, hash_ in self.container.loaded_files.items():
+            if exclude_extensions:
+                if os.path.splitext(fn)[1].lower() in exclude_extensions:
+                    continue
+
+            if hash_ not in hash_to_contents:
+                hash_to_contents[hash_] = self.dump_file(fn, hash_)
+            hash_to_file.setdefault(hash_, []).append(fn)
+        return result
+
+    async def dump_pv_relations(self) -> dict:
+        return apischema.serialize(
+            PVRelationshipResponse(
+                pv_relations=self.container.pv_relations,
+                script_relations=self.script_relations,
+                # can be generated on the client side easily with the cache
+                ioc_to_pvs={},
+            )
+        )
+
+    async def dump_pv_graphs(self) -> dict:
+        relation_graphs = {}
+        for record in self.container.database:
+            if record in self.container.pv_relations:
+                logger.info(
+                    "Generating graph for record %s which has %d relation(s)",
+                    record,
+                    len(self.container.pv_relations[record]),
+                )
+                graph = self.get_graph(
+                    (record,),
+                    graph_type="record",
+                    # format="svg",
+                )
+                # rendered = await self.get_graph_rendered(
+                #     (record,),
+                #     graph_type="record",
+                #     format="svg",
+                # )
+                relation_graphs[record] = {
+                    "source": graph.source,
+                    # "svg": rendered.decode("utf-8"),
+                }
+
+        return relation_graphs
+
+    def _prune_for_partial_dump(self) -> None:
+        # TODO: this shouldn't have to be destructive
+        logger.info("Pruning database definitions for a partial dump")
+        for ioc in self.container.scripts.values():
+            ioc.shell_state.database_definition = None
+
+    async def dump(self, full: bool = True) -> dict:
+        t0 = time.monotonic()
+
+        dumped = {
+            "plugins": self.dump_plugins(),
+            "pv_graphs": await self.dump_pv_graphs(),
+            "pv_relations": await self.dump_pv_relations(),
+        }
+
+        if not full:
+            self._prune_for_partial_dump()
+
+        dumped["iocs"] = self.dump_iocs()
+        dumped["files"] = self.dump_files(
+            exclude_extensions=[] if full else [".dbd"]
+        )
+
+        # Remove some duplicated information - PV relations are tracked
+        # at the top-level:
+        for item in dumped["iocs"].values():
+            item["ioc"].pop("pv_relations")
+
+        elapsed = time.monotonic() - t0
+        logger.info(
+            (
+                "Dumped %d plugins, %d files, %d record link graphs, "
+                "%d inter-IOC relationships, and %d IOCs in %.1f seconds"
+            ),
+            len(dumped["plugins"]),
+            len(dumped["files"]),
+            len(dumped["pv_graphs"]),
+            len(dumped["pv_relations"]["script_relations"]),
+            len(dumped["iocs"]),
+            elapsed,
+        )
+        return dumped
+
+    async def dump_to_file(
+        self,
+        filename: pathlib.Path,
+        indent: Optional[int] = None,
+        encoding: str = "utf-8",
+        full: bool = True,
+    ) -> None:
+        if filename.suffix.lower() not in (".json", ".gz"):
+            raise RuntimeError(
+                f"Output format not supported yet: {filename.suffix}. "
+                f"You probably want .json.gz, by the way, if you're intending "
+                f"to use the frontend with these results."
+            )
+
+        output = await self.dump(full=full)
+        dumped = json.dumps(output, sort_keys=True, indent=indent)
+        if filename.suffix.lower() in (".json", ):
+            with open(filename, "wt") as fp:
+                print(dumped, file=fp)
+        elif filename.suffix.lower() in (".gz", ):
+            with gzip.open(filename, "wb") as fp:
+                fp.write(dumped.encode(encoding))
+
     @property
     def iocs_by_name(self) -> Dict[str, IocMetadata]:
         """Dictionary of IOC name to IocMetadata."""
@@ -154,6 +375,22 @@ class ServerState:
         for plugin in self.plugins:
             self.tasks.create(self._update_plugin_loop(plugin))
 
+    async def _update_plugin(self, plugin: ServerPluginSpec):
+        """Update the specified plugin."""
+        with common.time_context() as ctx:
+            try:
+                await plugin.update()
+            except Exception:
+                logger.exception(
+                    "Failed to update plugin %r [%.1f s]",
+                    plugin.name, ctx()
+                )
+            else:
+                logger.info(
+                    "Successfully updated plugin %r [%.1f s]",
+                    plugin.name, ctx()
+                )
+
     async def _update_plugin_loop(self, plugin: ServerPluginSpec):
         while self.running and self.update_count == 0 and plugin.after_iocs:
             # Wait until IOCs have been loaded before updating this one for the
@@ -164,58 +401,51 @@ class ServerState:
 
         while self.running:
             logger.info("Updating plugin: %s", plugin.name)
-            with common.time_context() as ctx:
-                try:
-                    await plugin.update()
-                except Exception:
-                    logger.exception(
-                        "Failed to update plugin %r [%.1f s]",
-                        plugin.name, ctx()
-                    )
-                else:
-                    logger.info(
-                        "Successfully updated plugin %r [%.1f s]",
-                        plugin.name, ctx()
-                    )
-
+            await self._update_plugin(plugin)
             # for record, md in info["record_to_metadata"].items():
             #     ...
             await asyncio.sleep(settings.SERVER_SCAN_PERIOD)
 
         logger.info("Server plugin %r updates finished.", plugin.name)
 
+    async def _update_script_step(self) -> bool:
+        """Update scripts from the script loader."""
+        logger.info("Checking for new or updated IOCs...")
+        await self.update_script_loaders()
+
+        logger.info("Checking for changed scripts and database files...")
+        self._load_gateway_config()
+
+        updated = self.get_updated_iocs()
+        if not updated:
+            logger.info("No changes found.")
+            return False
+
+        logger.info(
+            "%d IOC%s changed.", len(updated),
+            " has" if len(updated) == 1 else "s have"
+        )
+        for idx, ioc in enumerate(updated[:10], 1):
+            logger.info("* %d: %s", idx, ioc.name)
+        if len(updated) > 10:
+            logger.info("... and %d more", len(updated) - 10)
+
+        with common.time_context() as ctx:
+            await self.update_iocs(updated)
+            logger.info(
+                "Updated %d IOCs in %.1f seconds", len(updated), ctx()
+            )
+        self.clear_cache()
+        return True
+
     async def _update_loop(self):
         """Update scripts from the script loader and watch for updates."""
         while self.running:
-            logger.info("Checking for new or updated IOCs...")
-            await self.update_script_loaders()
-
-            logger.info("Checking for changed scripts and database files...")
-            self._load_gateway_config()
-
-            updated = self.get_updated_iocs()
-            if not updated:
-                logger.info("No changes found.")
-                await asyncio.sleep(settings.SERVER_SCAN_PERIOD)
-                continue
-
-            logger.info(
-                "%d IOC%s changed.", len(updated),
-                " has" if len(updated) == 1 else "s have"
-            )
-            for idx, ioc in enumerate(updated[:10], 1):
-                logger.info("* %d: %s", idx, ioc.name)
-            if len(updated) > 10:
-                logger.info("... and %d more", len(updated) - 10)
-
-            with common.time_context() as ctx:
-                await self.update_iocs(updated)
-                logger.info(
-                    "Updated %d IOCs in %.1f seconds", len(updated), ctx()
-                )
-            self.clear_cache()
-
-            await asyncio.sleep(settings.SERVER_SCAN_PERIOD)
+            changes = await self._update_script_step()
+            poll_period = settings.SERVER_SCAN_PERIOD
+            if not changes:
+                poll_period *= 2
+            await asyncio.sleep(poll_period)
 
         logger.info("Server script updates finished.")
 
@@ -315,6 +545,36 @@ class ServerState:
                 logger.debug("New gateway file: %s (%s)", filename, pvlist.hash)
                 self.container.loaded_files[str(filename)] = pvlist.hash
 
+    def get_record_annnotations(
+        self,
+        ioc: LoadedIoc,
+        instance: RecordInstance,
+    ) -> Dict[StringWithContext, Any]:
+        """
+        Get IOC/plugin-provided annotations for the given record.
+        """
+        md = {}
+
+        if not instance.is_pva:
+            # For now, V3 only
+            gateway_key = StringWithContext("gateway", context=())
+            md[gateway_key] = apischema.serialize(
+                self.get_gateway_matches(instance.name)
+            )
+
+            md.update(ioc.shell_state.annotate_record(instance))
+
+        for plugin in self.plugins:
+            if not plugin.results:
+                continue
+
+            info = list(plugin.results.find_record_metadata(instance.name))
+            if info:
+                plugin_key = StringWithContext(plugin.name, context=())
+                instance.metadata[plugin_key] = info
+
+        return md
+
     def annotate_whatrec(self, ioc: LoadedIoc, what: WhatRecord) -> WhatRecord:
         """
         Annotate WhatRecord instances with things ServerState knows about.
@@ -327,22 +587,8 @@ class ServerState:
             if instance is None:
                 continue
 
-            if not instance.is_pva:
-                # For now, V3 only
-                instance.metadata["gateway"] = apischema.serialize(
-                    self.get_gateway_matches(instance.name)
-                )
-
-                ioc.shell_state.annotate_record(instance)
-
-            for plugin in self.plugins:
-                if not plugin.results:
-                    continue
-
-                info = list(plugin.results.find_record_metadata(instance.name))
-                if info:
-                    plugin_key = StringWithContext(plugin.name, context=())
-                    instance.metadata[plugin_key] = info
+            md = self.get_record_annnotations(ioc, instance)
+            instance.metadata.update(md)
 
         return what
 
@@ -593,6 +839,7 @@ class ServerHandler:
     async def api_pv_get_info(self, request: web.Request):
         pv_names = request.query.getall("pv")
         info = {pv_name: self.state.whatrec(pv_name) for pv_name in pv_names}
+        " {pv_name: ..., present: True, info=[whatrec]}"
         return serialized_response(
             {
                 pv_name: PVGetInfo(
@@ -717,10 +964,11 @@ class ServerHandler:
 
     @routes.get("/api/gateway/info")
     async def api_gateway_info(self, request: web.Request):
-        return serialized_response(self.state.gateway_config.pvlists or {})
+        return serialized_response(self.state.gateway_config)
 
     @routes.get("/api/file/info")
     async def api_ioc_info(self, request: web.Request):
+        # TODO: support new 'raw' response
         # script_name = pathlib.Path(request.query["file"])
         filename = request.query["file"]
         ioc_name = self.state.container.startup_script_to_ioc.get(filename, None)
@@ -933,7 +1181,9 @@ def main(
     standin_directory: Optional[Union[List, Dict]] = None,
     port: int = 8898,
     use_tracemalloc: bool = False,
+    offline_dump_target: Optional[AnyPath] = None,
     start: bool = True,
+    partial_dump: bool = False,
 ):
     if use_tracemalloc and tracemalloc is not None:
         tracemalloc.start()
@@ -948,7 +1198,18 @@ def main(
         standin_directory=standin_directory,
     )
     handler = ServerHandler(state)
-    app = _new_server(handler, port=port, run=True)
+    if offline_dump_target is None:
+        # Run the server.
+        run = True
+    else:
+        # Dump to a file - but don't start the server.
+        asyncio.run(state.trigger_manual_update(annotate_records=True))
+        asyncio.run(
+            state.dump_to_file(pathlib.Path(offline_dump_target), full=not partial_dump)
+        )
+        run = False
+
+    app = _new_server(handler, port=port, run=run)
 
     if use_tracemalloc and tracemalloc is not None:
         global tracemalloc_snapshot
